+++ model/csharp_writer.go
@@ -61,7 +61,7 @@ func (csw *CSharpWriter) GetField(_ int, field Field) string {
 
 func (csw *CSharpWriter) GetTypeName(typeName string) string {
 	switch typeName {
-	case "int":
+	case "int16":
 		return "int"
 	case "int32":
 		return "int"
